+++ gammapy/stats/__init__.py
@@ -4,6 +4,7 @@ Statistics utility functions and classes
 """
 from .significance import *
 from .data import *
+from .feldman_cousins import *
 from .fit_statistics import *
 from .poisson import *
 from .utils import *